Finish Feature Catgory whitelist
Description:
Checks if entered category is already in the database. If not it will prompt the user to add it to the white list or discard.
There is no extra saving of the categorys in the whitelist as only categorys which were accepted by the white list are allowed to be added.

Problems:
- After adding a category to the white list and pressing edit. The white list gets overwritten with the new edited entry, also if the user doesn't accept the white list prompt.
- Major handling of the white list happens in addentrywindow and therefore breaks the MVC-Pattern

the problems will be fixed in the next feature 'moveDialogLogic'