Move whole dialog logic to entrycontroller to keep MVC-Pattern
Description:
Before the logic behind the dialog windows (error and white list) were all split up in various classes where it didnt belong to. THis caused confusion and bugs could occur more easily.

Also changed the behavior of the entryForm when editing.
If you edit a entryForm now and you dont satisfy the white list or the entry correctness, the entryForm gets set back to its previous state.

This is implemented so the user can not add a entry wihtout adding it to the white list which was possible before