Add check for unsubmitted Edits
When clicking confirm, the entryWindow loops through all "EntryForms" and calls its checkForUnsubmittedEdit function which returns true, if an entry is in edit mode but hasnt been submitted.
This is implemented for security reasons so that the user doesnt lose any edits he made if he forgot to save them. In this case a error/note is raised to inform the user. He then can submit or delete the edit and proceed